Remove the no_newline_at_eol test
Turns out, this test was invalid. The root cause with this is that, in
essence, Kubernetes expects logs to consist of line, with line being
defined as in POSIX - a sequence of characters *ending with \n*.
Thus it's *not valid* to emit a log line without the terminating newline
symbol in Kubernetes.
One effect of this is that when using the CRI log format, lines won't be
considered complete until we emit a newline character arrives - and the
additional content before the newline will be added to the log line that's
missing the newline.

Given all of the above, there's no reason for this test to exist. The
reason it was added was the behaviour detail of the docker log driver,
but it's a mere implementation detail, and it we should abstract from it.

The original statement of the test is also ill-posed, cause, as explained
above, it's non-partial messages (and, generally speaking, any message)
that doesn't end with newline isn't a valid log line in the first place.

Signed-off-by: MOZGIII <mike-n@narod.ru>